Completed merge of MeasureAngleTool.
